% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDT_cov.R
\name{BSDT_cov}
\alias{BSDT_cov}
\title{Bayesian Standardised Difference Test with Covariates}
\usage{
BSDT_cov(
  case_tasks,
  case_covar,
  control_tasks,
  control_covar,
  alternative = c("two.sided", "greater", "less"),
  int.level = 0.95,
  calibrated = TRUE,
  iter = 1000
)
}
\arguments{
\item{case_tasks}{A vector of length 2. The case scores from the two tasks.}

\item{case_covar}{A vector containing the case scores on all covariates
included.}

\item{control_tasks}{A matrix or dataframe with 2 columns and n rows
containing the control scores for the two tasks.}

\item{control_covar}{A matrix or dataframe cointaining the control scores on
the covariates included.}

\item{alternative}{A character string specifying the alternative hypothesis,
must be one of \code{"two.sided"} (default), \code{"greater"} or
\code{"less"}. You can specify just the initial letter. Since the direction
of the expected effect depends on which task is set as X and which is set
as Y, be very careful if changing this parameter.}

\item{int.level}{The probability level on the Bayesian credible intervals.}

\item{calibrated}{Whether or not to use the standard theory (Jeffreys) prior
distribution (if set to \code{FALSE}) or a calibrated prior examined by
Berger and Sun (2008) and sample size treated as n - 1. See Crawford et al.
(2011) for further information. Calibrated prior is recommended.}

\item{iter}{Number of iterations to be performed. Greater number gives better
estimation but takes longer to calculate.}
}
\value{
A list with class \code{"htest"} containing the following components:
  \tabular{llll}{ \code{statistic}   \tab the average z-value over
  \code{iter} number of iterations. \cr\cr \code{p.value}    \tab the average
  p-value over \code{iter} number of iterations. \cr\cr \code{estimate} \tab
  case scores expressed as z-scores on task X and Y. Standardised effect size
  (Z-DCCC) of task difference between case and controls and point estimate of
  the proportion of the control population estimated to show a more extreme
  task difference. \cr\cr  \code{null.value}   \tab the value of the
  difference between tasks under the null hypothesis.\cr\cr \code{interval}
  \tab named numerical vector containing level of confidence and confidence
  intervals for both effect size and p-value.\cr\cr \code{desc}     \tab data
  frame containing means and standard deviations for controls as well as case
  scores. \cr\cr \code{cor.mat} \tab matrix giving the correlations between
  the tasks of interest and the covariates included. \cr\cr
  \code{sample.size}     \tab number of controls. Remember, if
  \code{calibrated} left as default (i.e. \code{TRUE}) sample size is treated
  as n - 1.\cr\cr \code{alternative}     \tab a character string describing
  the alternative hypothesis.\cr\cr \code{method} \tab a character string
  indicating what type of test was performed.\cr\cr \code{data.name} \tab a
  character string giving the name(s) of the data}
}
\description{
Bayesian Standardised Difference Test with Covariates
}
\examples{
  controls <- MASS::mvrnorm(20, mu = c(100, 80, 50),
   Sigma = matrix(c(15^2, 108, 78, 108, 10^2, 12,78, 12, 10^2),
   nrow = 3, byrow = TRUE), empirical = TRUE)

   BSDT_cov(c(70, 78), 30, controls[ , 1:2], controls[ , 3], iter = 1000)


}
