% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.R
\name{ch.ttest}
\alias{ch.ttest}
\title{Crawford and Howell's (1998) modified t-test}
\usage{
ch.ttest(
  case,
  controls,
  controls.sd = NULL,
  controls.n = NULL,
  alternative = c("less", "greater", "two.sided"),
  estimate.p = FALSE,
  conf.int = TRUE,
  conf.level = 0.95,
  conf.int.spec = 0.01,
  na.rm = FALSE
)
}
\arguments{
\item{case}{Case observation, can only be a single value.}

\item{controls}{Numeric vector of observations from the control sample.
If single value, treated as mean.}

\item{controls.sd}{If input of controls is single value, the standard deviation
of the sample must be gven as well.}

\item{controls.n}{If input of controls is single value, the size
of the sample must be gven as well.}

\item{alternative}{A character string specifying the alternative hypothesis,
must be one of \code{"less"} (default), \code{"greater"} or
\code{"two.sided"}. You can specify just the initial letter.}

\item{estimate.p}{Set to \code{TRUE} to estimate the proportion of population
falling above or below case score instead of standardised difference.}

\item{conf.int}{Calculate confidence intervals for desired estimate. Uses
iterative method, set to \code{FALSE} for faster calculation (e.g. for
simulations).}

\item{conf.level}{Level of confidence for intervals.}

\item{conf.int.spec}{The size of iterative steps for calculating confidence
intervals. Smaller values gives more precise intervals but takes longer to
calculate.}

\item{na.rm}{Remove \code{NA}s from controls.}
}
\value{
A list with class \code{"htest"} containing the following components:
\tabular{llll}{
\code{statistic}   \tab the value of the t-statistic.\cr\cr  \code{parameter}
\tab the degrees of freedom for the t-statistic.\cr\cr \code{p.value}    \tab
the p-value for the test.\cr\cr \code{estimate}    \tab estimated standardised
difference or proportion as well as mean and sd of controls.\cr\cr
\code{null.value}   \tab the value of the difference under the null
hypothesis.\cr\cr \code{sample.size}   \tab the size of the control sample\cr\cr
\code{conf.int}     \tab a confidence interval for the
specified estimate and alternative hypothesis.\cr\cr \code{interval}     \tab
string indicating the type of interval (proportion or difference). \cr\cr
\code{stderr}     \tab the standard error of the mean (difference), used as
denominator in the t-statistic formula.\cr\cr \code{alternative}     \tab a
character string describing the alternative hypothesis.\cr\cr \code{method}
\tab a character string indicating what type of t-test was performed.\cr\cr
\code{data.name}     \tab a character string giving the name(s) of the data
as well as sum summaries.
}
}
\description{
Takes a single observation and compares it to a distribution estimated by a
control sample. Calculates standardised difference between case and controls
or proportions falling above or below the case score as well as associated
confidence intervals.
}
\details{
Returns either the point estimate of the standardised difference
between the case score and the mean of the controls or the point estimate
of the p-value (i.e. the percentage of the population that would be
expected to obtain a lower or higher score, depending on the alternative
hypothesis).
}
\section{Note of caution}{

Calculating the confidence intervals relies on finding non-centrality
parameters for non-central t-distributions. Depending on the degrees of
freedom, the confidence level and the effect size exact accuracy from the
\code{stats::qt()} function used can not be guaranteed. However, the
approximations should be good enough for most cases.
See \url{https://stat.ethz.ch/pipermail/r-help/2008-June/164843.html}.
}

\examples{
ch.ttest(-2, rnorm(15), alternative = "l", estimate.p = TRUE)
}
