% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSDT.R
\name{RSDT}
\alias{RSDT}
\title{Revised Standardised Difference Test}
\usage{
RSDT(
  case.x,
  case.y,
  controls.x,
  controls.y,
  controls.x.sd = NULL,
  controls.y.sd = NULL,
  controls.n = NULL,
  cor.x.y = NULL,
  alpha = 0.05,
  exact.method = T
)
}
\arguments{
\item{case.x}{Case's score on task X.}

\item{case.y}{Case's score on task Y.}

\item{controls.x}{Controls' scores on task X. Takes either a vector of
observations or a single value interpreted as mean. \emph{Note}: you can supply
a vector as input for task X while mean and SD for task Y.}

\item{controls.y}{Controls' scores on task Y. Takes either a vector of
observations or a single value interpreted as mean. \emph{Note}: you can supply
a vector as input for task Y while mean and SD for task X.}

\item{controls.x.sd}{If single value for task X is given as input you must
supply the standard deviation of the sample.}

\item{controls.y.sd}{If single value for task Y is given as input you must
supply the standard deviation of the sample.}

\item{controls.n}{If X or Y is given as mean and SD you must supply the
sample size. If controls.x is given as vector and controls.y as mean and
SD, controls.n must equal the number of observations in controls.x.}

\item{cor.x.y}{If X or Y is given as mean and SD you must supply the
correlation between the tasks.}

\item{alpha}{Chosen risk of Type I errors.}

\item{exact.method}{Method for deriving the test statistic. Should rarely be
changed. See Crawford and Garthwaite (2005) for further information.}
}
\value{
A list with class \code{"htest"} containing the following components:
  \tabular{llll}{ \code{statistic}   \tab the value of the t-statistic.\cr\cr
  \code{parameter} \tab the degrees of freedom for the t-statistic.\cr\cr
  \code{p.value}    \tab the p-value for the test.\cr\cr \code{estimate}
  \tab case scores expressed as z-scores on task X and Y. Standardised effect
  size (Z-DCC) of task difference between case and controls and
  point estimate of the proportion of the control population estimated to show a
  more extreme task difference. \cr\cr \code{sample.size}   \tab the size of
  the control sample\cr\cr \code{null.value}   \tab the value of the
  difference under the null hypothesis.\cr\cr  \code{alternative}     \tab a
  character string describing the alternative hypothesis.\cr\cr \code{method}
  \tab a character string indicating what type of test was performed.\cr\cr
  \code{data.name}     \tab a character string giving the name(s) of the data}
}
\description{
A test used to determine dissociation between two functions as assessed by
two tasks.Takes two single values from a case (from two tasks) and compares
them to two a distribution of scores for both tasks estimated by a control
sample. Calculates a standardised effects size of task difference as well as
a point estimate of the proportion of the control population that would be
expected to show a more extreme task difference.
}
\examples{
RSDT(-3.857, -1.875, controls.x = 0, controls.y = 0, controls.x.sd = 1,
controls.y.sd = 1, controls.n = 20, cor.x.y = 0.68)
RSDT(-3.857, -1.875, controls.x = rnorm(20), controls.y = rnorm(20))
}
