% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singcar.R
\docType{package}
\name{singcar}
\alias{singcar}
\title{singcar: A package for analysing case comparisons to normative controls}
\description{
The singcar package provides functions for analysing single case data, mostly based on the work by Crawford and Garthwaite.
}
\section{singcar functions}{

\code{\link{ch.ttest}}
}

