% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UDT.R
\name{UDT}
\alias{UDT}
\title{Unstandardised Difference Test}
\usage{
UDT(
  case.x,
  case.y,
  controls.x,
  controls.y,
  controls.x.sd = NULL,
  controls.y.sd = NULL,
  controls.n = NULL,
  cor.x.y = NULL,
  alternative = c("two.sided", "greater", "less"),
  na.rm = FALSE
)
}
\arguments{
\item{case.x}{Case's score on task X.}

\item{case.y}{Case's score on task Y.}

\item{controls.x}{Controls' scores on task X. Takes either a vector of
observations or a single value interpreted as mean. \emph{Note}: you can
supply a vector as input for task X while mean and SD for task Y.}

\item{controls.y}{Controls' scores on task Y. Takes either a vector of
observations or a single value interpreted as mean. \emph{Note}: you can
supply a vector as input for task Y while mean and SD for task X.}

\item{controls.x.sd}{If single value for task X is given as input you must
supply the standard deviation of the sample.}

\item{controls.y.sd}{If single value for task Y is given as input you must
supply the standard deviation of the sample.}

\item{controls.n}{If X or Y is given as mean and SD you must supply the
sample size. If controls.x is given as vector and controls.y as mean and
SD, controls.n must equal the number of observations in controls.x.}

\item{cor.x.y}{If X or Y is given as mean and SD you must supply the
correlation between the tasks.}

\item{alternative}{A character string specifying the alternative hypothesis,
must be one of \code{"two.sided"} (default), \code{"greater"} or
\code{"less"}. You can specify just the initial letter. Since the direction
of the expected effect depends on which task is set as X and which is set
as Y, be very careful if changing this parameter.}

\item{na.rm}{Remove \code{NA}s from controls.}
}
\value{
A list with class \code{"htest"} containing the following components:
  \tabular{llll}{ \code{statistic}   \tab the t-statistic. \cr\cr
  \code{parameter} \tab the degrees of freedom for the t-statistic.\cr\cr
  \code{p.value}    \tab the p-value of the test.\cr\cr \code{estimate} \tab
  unstandardised case scores, task difference and pont estimate of proportion
  control population expected to above or below the observed task difference.
  \cr\cr \code{control.desc}   \tab named numerical with descriptive
  statistics of the control samples. \cr\cr \code{null.value}   \tab the
  value of the difference under the null hypothesis.\cr\cr
  \code{alternative}     \tab a character string describing the alternative
  hypothesis.\cr\cr \code{method} \tab a character string indicating what
  type of test was performed.\cr\cr \code{data.name} \tab a character string
  giving the name(s) of the data}
}
\description{
A test on the difference between two tasks in a single case, by comparison to
a control sample. Use only when the two tasks are measured on the \emph{same}
scale.
}
\examples{
UDT(-3.857, -1.875, controls.x = 0, controls.y = 0, controls.x.sd = 1,
controls.y.sd = 1, controls.n = 20, cor.x.y = 0.68)
UDT(-3.857, -1.875, controls.x = rnorm(20), controls.y = rnorm(20))

}
\references{
{Crawford, J. R., & Garthwaite, P. H. (2005). Testing for
Suspected Impairments and Dissociations in Single-Case Studies in
Neuropsychology: Evaluation of Alternatives Using Monte Carlo Simulations and
Revised Tests for Dissociations. \emph{Neuropsychology, 19}(3), 318 - 331.
\url{https://doi.org/10.1037/0894-4105.19.3.318}}
}
