% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDT.R
\name{BSDT}
\alias{BSDT}
\title{Bayesian Standardised Difference Test}
\usage{
BSDT(
  case_a,
  case_b,
  controls_a,
  controls_b,
  sd_a = NULL,
  sd_b = NULL,
  sample_size = NULL,
  r_ab = NULL,
  alternative = c("two.sided", "greater", "less"),
  int_level = 0.95,
  iter = 1000,
  unstandardised = FALSE,
  calibrated = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{case_a}{Case's score on task A.}

\item{case_b}{Case's score on task B.}

\item{controls_a}{Controls' scores on task A. Takes either a vector of
observations or a single value interpreted as mean. \emph{Note}: you can
supply a vector as input for task A while mean and SD for task B.}

\item{controls_b}{Controls' scores on task A. Takes either a vector of
observations or a single value interpreted as mean. \emph{Note}: you can
supply a vector as input for task B while mean and SD for task A.}

\item{sd_a}{If single value for task A is given as input you must
supply the standard deviation of the sample.}

\item{sd_b}{If single value for task B is given as input you must
supply the standard deviation of the sample.}

\item{sample_size}{If A or B is given as mean and SD you must supply the
sample size. If controls_a is given as vector and controls_b as mean and
SD, sample_size must equal the number of observations in controls_a.}

\item{r_ab}{If A or B is given as mean and SD you must supply the
correlation between the tasks.}

\item{alternative}{A character string specifying the alternative hypothesis,
must be one of \code{"two.sided"} (default), \code{"greater"} or
\code{"less"}. You can specify just the initial letter. Since the direction
of the expected effect depends on which task is set as A and which is set
as B, be very careful if changing this parameter.}

\item{int_level}{Level of confidence for credible intervals.}

\item{iter}{Number of iterations. Greater number gives better
estimation but takes longer to calculate.}

\item{unstandardised}{Estimate z-value based on standardised or
unstandardised task scores.}

\item{calibrated}{set to \code{TRUE} to use a calibrated prior distribution.}

\item{na.rm}{Remove \code{NA}s from controls.}
}
\value{
A list with class \code{"htest"} containing the following components:
  \tabular{llll}{ \code{statistic}   \tab the average z-value over
  \code{iter} number of iterations. \cr\cr \code{p.value}    \tab the average
  p-value over \code{iter} number of iterations. \cr\cr \code{estimate} \tab
  case scores expressed as z-scores on task A and B. Standardised effect size
  (Z-DCC) of task difference between case and controls and point estimate of
  the proportion of the control population estimated to show a more extreme
  task difference. \cr\cr  \code{null.value}   \tab the value of the
  difference under the null hypothesis.\cr\cr  \code{alternative}     \tab a
  character string describing the alternative hypothesis.\cr\cr \code{method}
  \tab a character string indicating what type of test was performed.\cr\cr
  \code{data.name} \tab a character string giving the name(s) of the data}
}
\description{
A test on the difference between two tasks in a single case, by comparison to
the difference of the same two tasks in a control sample. Calculates a
standardised effects size of task difference as well as a point estimate of
the proportion of the control population that would be expected to show a
more extreme task difference.
}
\examples{
BSDT(-3.857, -1.875, controls_a = 0, controls_b = 0, sd_a = 1,
sd_b = 1, sample_size = 20, r_ab = 0.68)

BSDT(case_a = size_weight_illusion[1, "V_SWI"], case_b = size_weight_illusion[1, "K_SWI"],
 controls_a = size_weight_illusion[-1, "V_SWI"], controls_b = size_weight_illusion[-1, "K_SWI"])

}
\references{
Crawford, J. R., & Garthwaite, P. H. (2007). Comparison of a single case to a
control or normative sample in neuropsychology: Development of a Bayesian
approach. \emph{Cognitive Neuropsychology, 24}(4), 343–372.
\url{https://doi.org/10.1080/02643290701290146}
}
